refactor(server): split server-only modules into .server.ts files
Move DB-touching code (auth helpers, plan helpers, plan cleanup, polar
handlers, custom domain loader, field chunk manifest) into `.server.ts`
companions so the bundler keeps `pg`, drizzle, and the Polar SDK out of
the client graph. Pure isomorphic helpers stay in the original `.ts`
files. Replace the eager drizzle init in `@/db` with a lazy proxy that
loads `drizzle-orm/node-postgres` via `createRequire` only on first
access, and convert `formProSettingsMiddleware` to compose off
`authMiddleware` for properly-typed `context.session`.
