+++ src/lib/auth/middleware.ts
@@ -1,12 +1,17 @@
 import { redirect } from "@tanstack/react-router";
 import { createMiddleware } from "@tanstack/react-start";
 import { getCookie, getRequestHeaders, getRequestUrl } from "@tanstack/react-start/server";
-import { auth } from "@/lib/auth/auth";
 import { getActiveOrgId } from "@/lib/server-fn/auth-helpers";
-import { getOrgPlan, requiresProForFormSettings } from "@/lib/server-fn/plan-helpers";
+import { requiresProForFormSettings } from "@/lib/server-fn/plan-helpers";
 import type { FormProSettingsInput } from "@/lib/server-fn/plan-helpers";
 
+// `auth` is lazy-imported inside the server body. A static import here would
+// drag the entire `@polar-sh/sdk` + `@/db` + `pg` graph into the client
+// bundle — this file is statically imported by every server-fn module
+// (`forms.ts`, `workspaces.ts`, …) which are in turn imported by route
+// components, putting middleware.ts squarely in the client graph.
 export const authMiddleware = createMiddleware().server(async ({ next }) => {
+  const { auth } = await import("@/lib/auth/auth");
   const headers = getRequestHeaders();
   const session = await auth.api.getSession({ headers });
 
@@ -26,25 +31,22 @@ export const authMiddleware = createMiddleware().server(async ({ next }) => {
   });
 });
 
-// Gates Pro-only fields on createForm / updateForm. Runs after authMiddleware
-// so the session is already validated. Reads `data` (validated input) and
-// throws before the handler runs when a free org tries to enable any
-// Pro-gated setting.
-export const formProSettingsMiddleware = createMiddleware({ type: "function" }).server(
-  async ({ next, data, context }) => {
+// Gates Pro-only fields on createForm / updateForm. `getOrgPlan` is dynamically
+// imported so the `@/db` graph never lands in client bundles that pick up this
+// file via `forms.ts` → route components.
+export const formProSettingsMiddleware = createMiddleware({ type: "function" })
+  .middleware([authMiddleware])
+  .server(async ({ next, data, context }) => {
     const input = data as unknown as FormProSettingsInput;
     if (!requiresProForFormSettings(input)) return next();
 
-    const session = (context as unknown as { session?: unknown }).session as Parameters<
-      typeof getActiveOrgId
-    >[0];
-    const plan = await getOrgPlan(getActiveOrgId(session));
+    const { getOrgPlan } = await import("@/lib/server-fn/plan-helpers.server");
+    const plan = await getOrgPlan(getActiveOrgId(context.session));
     if (plan === "free") {
       throw new Error("This feature requires a Pro subscription. Please upgrade to continue.");
     }
     return next();
-  },
-);
+  });
 
 export const guestMiddleware = createMiddleware().server(async ({ next }) => {
   // O(1) cookie check instead of DB round-trip (~1.97s saving)
