+++ src/lib/config/app-config.ts
@@ -1,7 +1,5 @@
 export const APP_NAME = import.meta.env.VITE_APP_NAME || "Reform";
 export const APP_WEBSITE_URL = import.meta.env.VITE_APP_WEBSITE_URL || "https://betterforms.com";
-export const APP_API_DOCS_URL =
-  import.meta.env.VITE_APP_API_DOCS_URL || "https://developers.betterforms.so";
 
 /** Default icon identifier used throughout the app (forms, pickers, previews) */
 export const DEFAULT_ICON = "default-icon";
