+++ src/lib/server-fn/org.ts
@@ -1,13 +1,18 @@
 import { queryOptions } from "@tanstack/react-query";
 import { createServerFn } from "@tanstack/react-start";
 import { getRequestHeaders } from "@tanstack/react-start/server";
-import { auth } from "@/lib/auth/auth";
+
 /**
  * Server-only: fetches org data with request headers for SSR.
  * Used by _authenticated loader - must stay in server fn to preserve client boundary.
  * Auth is handled by the route's authMiddleware — no need to re-validate session here.
+ *
+ * `auth` is lazy-imported inside the handler body. Static-importing it here
+ * would drag `@polar-sh/sdk` + `@/db` + `pg` into the client bundle (this
+ * file is statically imported by `_authenticated.tsx`).
  */
 export const getOrgDataForLayout = createServerFn({ method: "GET" }).handler(async () => {
+  const { auth } = await import("@/lib/auth/auth");
   const headers = getRequestHeaders();
   const [activeOrg, orgsData] = await Promise.all([
     auth.api.getFullOrganization({ headers }),
