+++ src/lib/server-fn/plan-helpers.ts
@@ -1,21 +1,7 @@
-import { eq } from "drizzle-orm";
-import { db } from "@/db";
-import { organization } from "@/db/schema";
-
 export type ServerPlan = "free" | "pro" | "biz";
 
-const isServerPlan = (value: unknown): value is ServerPlan =>
+export const isServerPlan = (value: unknown): value is ServerPlan =>
   value === "free" || value === "pro" || value === "biz";
-
-// Reads the cached `organization.plan` (synced by Polar webhooks); falls back
-// to 'free' for unknown orgs or unexpected column values.
-export const getOrgPlan = async (orgId: string): Promise<ServerPlan> => {
-  const [row] = await db
-    .select({ plan: organization.plan })
-    .from(organization)
-    .where(eq(organization.id, orgId));
-  return isServerPlan(row?.plan) ? row.plan : "free";
-};
 
 export type FormProSettingsInput = {
   branding?: boolean;
