+++ src/test/plan-cleanup.test.ts
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { eq } from "drizzle-orm";
 import { db } from "@/db";
 import { customDomains, forms, organization } from "@/db/schema";
-import { applyDowngradeCleanup, applyUpgradeRestore } from "@/lib/server-fn/plan-cleanup";
+import { applyDowngradeCleanup, applyUpgradeRestore } from "@/lib/server-fn/plan-cleanup.server";
 import {
   cleanupTestOrg,
   cleanupTestUser,
